feat: better logging for Stashy backends
The _init() function lifecycle now includes the logger and a guaranteed reference to the parent Stashy instance.

This can be used for debugging the initialization process itself.
